+++ lib/flutter_map_tile_caching.dart
@@ -1,2 +1,12 @@
-export 'src/storage_caching_tile_provider.dart';
-export 'src/tile_storage_caching_manager.dart';
+// This is flutter_map_tile_caching's main file
+// You should never need to include anything other than this file from this library in your project
+
+export 'src/regions/circle.dart';
+export 'src/regions/downloadableRegion.dart' show RegionType, DownloadProgress;
+export 'src/regions/rectangle.dart';
+
+export 'src/shapeChooser.dart';
+export 'src/general.dart';
+
+export 'src/main.dart';
+export 'src/databaseManager.dart';